+++ crates/r/R/extendr-wrappers.R
@@ -18,7 +18,7 @@ koon_browsers <- function() .Call(wrap__koon_browsers)
 
 Koon <- new.env(parent = emptyenv())
 
-Koon$new <- function(browser, proxy = NULL, proxies = NULL, timeout = NULL, randomize = NULL, headers = NULL, local_address = NULL, on_request = NULL, on_response = NULL, on_redirect = NULL, retries = NULL) .Call(wrap__Koon__new, browser, proxy, proxies, timeout, randomize, headers, local_address, on_request, on_response, on_redirect, retries)
+Koon$new <- function(browser, proxy = NULL, proxies = NULL, timeout = NULL, randomize = NULL, headers = NULL, local_address = NULL, on_request = NULL, on_response = NULL, on_redirect = NULL, retries = NULL, locale = NULL, proxy_headers = NULL, ip_version = NULL, follow_redirects = NULL, max_redirects = NULL, cookie_jar = NULL, session_resumption = NULL, ignore_tls_errors = NULL, doh = NULL) .Call(wrap__Koon__new, browser, proxy, proxies, timeout, randomize, headers, local_address, on_request, on_response, on_redirect, retries, locale, proxy_headers, ip_version, follow_redirects, max_redirects, cookie_jar, session_resumption, ignore_tls_errors, doh)
 
 Koon$get <- function(url, headers = NULL) .Call(wrap__Koon__get, self, url, headers)
 
@@ -43,6 +43,8 @@ Koon$total_bytes_sent <- function() .Call(wrap__Koon__total_bytes_sent, self)
 Koon$total_bytes_received <- function() .Call(wrap__Koon__total_bytes_received, self)
 
 Koon$reset_counters <- function() invisible(.Call(wrap__Koon__reset_counters, self))
+
+Koon$user_agent <- function() .Call(wrap__Koon__user_agent, self)
 
 Koon$clear_cookies <- function() invisible(.Call(wrap__Koon__clear_cookies, self))